requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.0.0
llama-index>=0.9.0
openai>=0.27.0
